+++ lib/cisco_node_utils/interface.rb
@@ -34,12 +34,24 @@ module Cisco
 
   # Interface - node utility class for general interface config management
   class Interface < NodeUtil
+    # Regexp to match various Ethernet interface variants:
+    #                       Ethernet
+    #                GigabitEthernet
+    #                             TenGigE
+    #                         HundredGigE
+    #                                     MgmtEth
+    ETHERNET = Regexp.new('(Ethernet|GigE|MgmtEth)', Regexp::IGNORECASE)
+    # Regexp to match various link bundle interface variants
+    PORTCHANNEL = Regexp.new('(port-channel|Bundle-Ether)', Regexp::IGNORECASE)
+
     attr_reader :name
 
     def initialize(name, instantiate=true)
       fail TypeError unless name.is_a?(String)
       fail ArgumentError unless name.length > 0
       @name = name.downcase
+      @smr = config_get('interface', 'stp_mst_range')
+      @svr = config_get('interface', 'stp_vlan_range')
 
       create if instantiate
     end
@@ -58,103 +70,38 @@ module Cisco
 
     def create
       feature_vlan_set(true) if @name[/vlan/i]
-      config_set('interface', 'create', @name)
+      config_set('interface', 'create', name: @name)
+    rescue Cisco::CliError
+      # Some XR platforms do not support channel-group configuration
+      # on some OS versions. Since this is an OS version difference and not
+      # a platform difference, we can't handle this in the YAML.
+      raise unless PORTCHANNEL =~ @name && platform == :ios_xr
+      raise Cisco::UnsupportedError.new('interface', @name, 'create')
     end
 
     def destroy
-      config_set('interface', 'destroy', @name)
+      config_set('interface', 'destroy', name: @name)
     end
 
     ########################################################
     #                      PROPERTIES                      #
     ########################################################
 
+    # For range based attributes, a new attribute purge will
+    # be added in future. When purge is set to true, all the
+    # ranges which are specified in the manifest will be set
+    # to the desired values and those which are not specified
+    # in the manifest will be set to default. When purge is
+    # false, only the ranges specified in the manifest will
+    # be set to the values given in the manifest and others
+    # are left untouched.
+
     def access_vlan
-      config_get('interface', 'access_vlan', @name)
+      config_get('interface', 'access_vlan', name: @name)
     end
 
     def access_vlan=(vlan)
-      config_set('interface', 'access_vlan', @name, vlan)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
-    end
-
-    def ipv4_acl_in
-      config_get('interface', 'ipv4_acl_in', @name)
-    end
-
-    def ipv4_acl_in=(val)
-      if val != ''
-        state = ''
-      else
-        state = 'no'
-        val = ipv4_acl_in
-      end
-
-      return unless val && val != ''
-      config_set('interface', 'ipv4_acl_in', @name, state, val)
-    end
-
-    def default_ipv4_acl_in
-      config_get_default('interface', 'ipv4_acl_in')
-    end
-
-    def ipv4_acl_out
-      config_get('interface', 'ipv4_acl_out', @name)
-    end
-
-    def ipv4_acl_out=(val)
-      if val != ''
-        state = ''
-      else
-        state = 'no'
-        val = ipv4_acl_out
-      end
-
-      return unless val && val != ''
-      config_set('interface', 'ipv4_acl_out', @name, state, val)
-    end
-
-    def default_ipv4_acl_out
-      config_get_default('interface', 'ipv4_acl_out')
-    end
-
-    def ipv6_acl_in
-      config_get('interface', 'ipv6_acl_in', @name)
-    end
-
-    def ipv6_acl_in=(val)
-      if val != ''
-        state = ''
-      else
-        state = 'no'
-        val = ipv6_acl_in
-      end
-      return unless val && val != ''
-      config_set('interface', 'ipv6_acl_in', @name, state, val)
-    end
-
-    def default_ipv6_acl_in
-      config_get_default('interface', 'ipv6_acl_in')
-    end
-
-    def ipv6_acl_out
-      config_get('interface', 'ipv6_acl_out', @name)
-    end
-
-    def ipv6_acl_out=(val)
-      if val != ''
-        state = ''
-      else
-        state = 'no'
-        val = ipv6_acl_out
-      end
-      return unless val && val != ''
-      config_set('interface', 'ipv6_acl_out', @name, state, val)
-    end
-
-    def default_ipv6_acl_out
-      config_get_default('interface', 'ipv6_acl_out')
+      config_set('interface', 'access_vlan', name: @name, vlan: vlan)
     end
 
     def default_access_vlan
@@ -162,18 +109,18 @@ module Cisco
     end
 
     def description
-      config_get('interface', 'description', @name)
+      config_get('interface', 'description', name: @name)
     end
 
     def description=(desc)
       fail TypeError unless desc.is_a?(String)
       if desc.strip.empty?
-        config_set('interface', 'description', @name, 'no', '')
+        config_set('interface', 'description',
+                   name: @name, state: 'no', desc: '')
       else
-        config_set('interface', 'description', @name, '', desc)
+        config_set('interface', 'description',
+                   name: @name, state: '', desc: desc)
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
     end
 
     def default_description
@@ -181,17 +128,17 @@ module Cisco
     end
 
     def encapsulation_dot1q
-      config_get('interface', 'encapsulation_dot1q', @name)
+      config_get('interface', 'encapsulation_dot1q', name: @name)
     end
 
     def encapsulation_dot1q=(val)
       if val.to_s.empty?
-        config_set('interface', 'encapsulation_dot1q', @name, 'no', '')
+        config_set('interface', 'encapsulation_dot1q',
+                   name: @name, state: 'no', vlan: '')
       else
-        config_set('interface', 'encapsulation_dot1q', @name, '', val)
+        config_set('interface', 'encapsulation_dot1q',
+                   name: @name, state: '', vlan: val)
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
     end
 
     def default_encapsulation_dot1q
@@ -207,13 +154,14 @@ module Cisco
     end
 
     def fabric_forwarding_anycast_gateway
-      config_get('interface', 'fabric_forwarding_anycast_gateway', @name)
+      config_get('interface', 'fabric_forwarding_anycast_gateway', name: @name)
     end
 
     def fabric_forwarding_anycast_gateway=(state)
       no_cmd = (state ? '' : 'no')
       config_set('interface',
-                 'fabric_forwarding_anycast_gateway', @name, no_cmd)
+                 'fabric_forwarding_anycast_gateway',
+                 name: @name, state: no_cmd)
       fail if fabric_forwarding_anycast_gateway.to_s != state.to_s
     rescue Cisco::CliError => e
       info = "[#{@name}] '#{e.command}' : #{e.clierror}"
@@ -254,8 +202,48 @@ module Cisco
         config_set('fex', 'feature', 'no') if curr == :enabled
         config_set('fex', 'feature_install', 'no')
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+    end
+
+    def ipv4_acl_in
+      config_get('interface', 'ipv4_acl_in', name: @name)
+    end
+
+    def ipv4_acl_in=(val)
+      if val != ''
+        state = ''
+      else
+        state = 'no'
+        val = ipv4_acl_in
+      end
+
+      return unless val && val != ''
+      config_set('interface', 'ipv4_acl_in',
+                 name: @name, state: state, acl: val)
+    end
+
+    def default_ipv4_acl_in
+      config_get_default('interface', 'ipv4_acl_in')
+    end
+
+    def ipv4_acl_out
+      config_get('interface', 'ipv4_acl_out', name: @name)
+    end
+
+    def ipv4_acl_out=(val)
+      if val != ''
+        state = ''
+      else
+        state = 'no'
+        val = ipv4_acl_out
+      end
+
+      return unless val && val != ''
+      config_set('interface', 'ipv4_acl_out',
+                 name: @name, state: state, acl: val)
+    end
+
+    def default_ipv4_acl_out
+      config_get_default('interface', 'ipv4_acl_out')
     end
 
     def ipv4_addr_mask_set(addr, mask, secondary=false)
@@ -275,13 +263,22 @@ module Cisco
         state = ''
         am = "#{addr}/#{mask}"
       end
-      config_set('interface', 'ipv4_addr_mask', @name, state, am, sec)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', 'ipv4_addr_mask',
+                 name: @name, state: state, addr: am, secondary: sec)
     end
 
     def ipv4_addr_mask
-      config_get('interface', 'ipv4_addr_mask', @name)
+      val = config_get('interface', 'ipv4_addr_mask', name: @name)
+      if val && platform == :ios_xr
+        # IOS XR reports address as <address> <bitmask> [secondary] but we
+        # want <address>/<length> [secondary]
+        val.each_with_index do |entry, i|
+          mask = entry[1].split(' ')
+          mask[0] = Utils.bitmask_to_length(mask[0])
+          val[i][1] = mask.join(' ')
+        end
+      end
+      val
     end
 
     def select_ipv4_attribute(attribute)
@@ -346,14 +343,15 @@ module Cisco
     end
 
     def ipv4_arp_timeout
-      config_get('interface', ipv4_arp_timeout_lookup_string, @name)
+      config_get('interface', ipv4_arp_timeout_lookup_string, name: @name)
     end
 
     def ipv4_arp_timeout=(timeout)
       fail "'ipv4 arp timeout' can ony be configured on a vlan interface" unless
         /vlan/.match(@name)
       state = (timeout == default_ipv4_arp_timeout) ? 'no' : ''
-      config_set('interface', 'ipv4_arp_timeout', @name, state, timeout)
+      config_set('interface', 'ipv4_arp_timeout',
+                 name: @name, state: state, timeout: timeout)
     end
 
     def default_ipv4_arp_timeout
@@ -361,16 +359,14 @@ module Cisco
     end
 
     def ipv4_pim_sparse_mode
-      config_get('interface', 'ipv4_pim_sparse_mode', @name)
+      config_get('interface', 'ipv4_pim_sparse_mode', name: @name)
     end
 
     def ipv4_pim_sparse_mode=(state)
       check_switchport_disabled
-      Pim.feature_enable unless Pim.feature_enabled
-      config_set('interface', 'ipv4_pim_sparse_mode', @name,
-                 state ? '' : 'no')
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      Pim.feature_enable unless platform == :ios_xr || Pim.feature_enabled
+      config_set('interface', 'ipv4_pim_sparse_mode',
+                 name: @name, state: state ? '' : 'no')
     end
 
     def default_ipv4_pim_sparse_mode
@@ -378,13 +374,13 @@ module Cisco
     end
 
     def ipv4_proxy_arp
-      config_get('interface', 'ipv4_proxy_arp', @name)
+      config_get('interface', 'ipv4_proxy_arp', name: @name)
     end
 
     def ipv4_proxy_arp=(proxy_arp)
       check_switchport_disabled
       no_cmd = (proxy_arp ? '' : 'no')
-      config_set('interface', 'ipv4_proxy_arp', @name, no_cmd)
+      config_set('interface', 'ipv4_proxy_arp', name: @name, state: no_cmd)
     end
 
     def default_ipv4_proxy_arp
@@ -401,17 +397,58 @@ module Cisco
     end
 
     def ipv4_redirects
-      config_get('interface', ipv4_redirects_lookup_string, @name)
+      config_get('interface', ipv4_redirects_lookup_string, name: @name)
     end
 
     def ipv4_redirects=(redirects)
       check_switchport_disabled
       no_cmd = (redirects ? '' : 'no')
-      config_set('interface', ipv4_redirects_lookup_string, @name, no_cmd)
+      config_set('interface', ipv4_redirects_lookup_string,
+                 name: @name, state: no_cmd)
     end
 
     def default_ipv4_redirects
       config_get_default('interface', ipv4_redirects_lookup_string)
+    end
+
+    def ipv6_acl_in
+      config_get('interface', 'ipv6_acl_in', name: @name)
+    end
+
+    def ipv6_acl_in=(val)
+      if val != ''
+        state = ''
+      else
+        state = 'no'
+        val = ipv6_acl_in
+      end
+      return unless val && val != ''
+      config_set('interface', 'ipv6_acl_in',
+                 name: @name, state: state, acl: val)
+    end
+
+    def default_ipv6_acl_in
+      config_get_default('interface', 'ipv6_acl_in')
+    end
+
+    def ipv6_acl_out
+      config_get('interface', 'ipv6_acl_out', name: @name)
+    end
+
+    def ipv6_acl_out=(val)
+      if val != ''
+        state = ''
+      else
+        state = 'no'
+        val = ipv6_acl_out
+      end
+      return unless val && val != ''
+      config_set('interface', 'ipv6_acl_out',
+                 name: @name, state: state, acl: val)
+    end
+
+    def default_ipv6_acl_out
+      config_get_default('interface', 'ipv6_acl_out')
     end
 
     def feature_lacp?
@@ -420,35 +457,41 @@ module Cisco
 
     def feature_lacp_set(val)
       return if feature_lacp? == val
-      config_set('interface', 'feature_lacp', val ? '' : 'no')
+      config_set('interface', 'feature_lacp', state: val ? '' : 'no')
+    end
+
+    def mtu_lookup_string
+      case @name
+      when /loopback/i
+        return 'mtu_loopback'
+      else
+        return 'mtu_other_interfaces'
+      end
     end
 
     def mtu
-      config_get('interface', 'mtu', @name)
+      config_get('interface', mtu_lookup_string, name: @name)
     end
 
     def mtu=(val)
       check_switchport_disabled
-      config_set('interface', 'mtu', @name, '', val)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', mtu_lookup_string,
+                 name: @name, state: '', mtu: val)
     end
 
     def default_mtu
-      config_get_default('interface', 'mtu')
+      config_get_default('interface', mtu_lookup_string)
     end
 
     def speed
-      config_get('interface', 'speed', @name)
+      config_get('interface', 'speed', name: @name)
     end
 
     def speed=(val)
       if node.product_id =~ /C31\d\d/
         fail 'Changing interface speed is not permitted on this platform'
       end
-      config_set('interface', 'speed', @name, val)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', 'speed', name: @name, speed: val)
     end
 
     def default_speed
@@ -456,16 +499,14 @@ module Cisco
     end
 
     def duplex
-      config_get('interface', 'duplex', @name)
+      config_get('interface', 'duplex', name: @name)
     end
 
     def duplex=(val)
       if node.product_id =~ /C31\d\d/
         fail 'Changing interface duplex is not permitted on this platform'
       end
-      config_set('interface', 'duplex', @name, val)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', 'duplex', name: @name, duplex: val)
     end
 
     def default_duplex
@@ -474,9 +515,9 @@ module Cisco
 
     def negotiate_auto_lookup_string
       case @name
-      when /Ethernet/i
+      when ETHERNET
         return 'negotiate_auto_ethernet'
-      when /port-channel/i # Ether-channel
+      when PORTCHANNEL
         return 'negotiate_auto_portchannel'
       else
         return 'negotiate_auto_other_interfaces'
@@ -484,13 +525,13 @@ module Cisco
     end
 
     def negotiate_auto
-      config_get('interface', negotiate_auto_lookup_string, @name)
+      config_get('interface', negotiate_auto_lookup_string, name: @name)
     end
 
     def negotiate_auto=(negotiate_auto)
       lookup = negotiate_auto_lookup_string
       no_cmd = (negotiate_auto ? '' : 'no')
-      config_set('interface', lookup, @name, no_cmd)
+      config_set('interface', lookup, name: @name, state: no_cmd)
     end
 
     def default_negotiate_auto
@@ -498,19 +539,17 @@ module Cisco
     end
 
     def shutdown
-      config_get('interface', 'shutdown', @name)
+      config_get('interface', 'shutdown', name: @name)
     end
 
     def shutdown=(state)
       no_cmd = (state ? '' : 'no')
-      config_set('interface', 'shutdown', @name, no_cmd)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', 'shutdown', name: @name, state: no_cmd)
     end
 
     def default_shutdown
       case @name
-      when /Ethernet/i
+      when ETHERNET
         def_sw = system_default_switchport
         def_shut = system_default_switchport_shutdown
 
@@ -529,7 +568,7 @@ module Cisco
       when /loopback/i
         lookup = 'shutdown_loopback'
 
-      when /port-channel/i # EtherChannel
+      when PORTCHANNEL
         lookup = 'shutdown_ether_channel'
 
       when /Vlan/i
@@ -541,28 +580,227 @@ module Cisco
       config_get_default('interface', lookup)
     end
 
+    def stp_bpdufilter
+      config_get('interface', 'stp_bpdufilter', name: @name)
+    end
+
+    def stp_bpdufilter=(val)
+      if val
+        state = ''
+      else
+        state = 'no'
+        val = ''
+      end
+      config_set('interface',
+                 'stp_bpdufilter', name: @name, state: state, filter: val)
+    end
+
+    def default_stp_bpdufilter
+      config_get_default('interface', 'stp_bpdufilter')
+    end
+
+    def stp_bpduguard
+      config_get('interface', 'stp_bpduguard', name: @name)
+    end
+
+    def stp_bpduguard=(val)
+      if val
+        state = ''
+      else
+        state = 'no'
+        val = ''
+      end
+      config_set('interface',
+                 'stp_bpduguard', name: @name, state: state, guard: val)
+    end
+
+    def default_stp_bpduguard
+      config_get_default('interface', 'stp_bpduguard')
+    end
+
+    def stp_cost
+      cost = config_get('interface', 'stp_cost', name: @name)
+      cost == 'auto' ? cost : cost.to_i
+    end
+
+    def stp_cost=(val)
+      config_set('interface', 'stp_cost', name: @name, cost: val)
+    end
+
+    def default_stp_cost
+      config_get_default('interface', 'stp_cost')
+    end
+
+    def stp_guard
+      config_get('interface', 'stp_guard', name: @name)
+    end
+
+    def stp_guard=(val)
+      if val
+        state = ''
+      else
+        state = 'no'
+        val = ''
+      end
+      config_set('interface', 'stp_guard', name: @name, state: state,
+                 guard: val)
+    end
+
+    def default_stp_guard
+      config_get_default('interface', 'stp_guard')
+    end
+
+    def stp_link_type
+      config_get('interface', 'stp_link_type', name: @name)
+    end
+
+    def stp_link_type=(val)
+      config_set('interface', 'stp_link_type', name: @name, type: val)
+    end
+
+    def default_stp_link_type
+      config_get_default('interface', 'stp_link_type')
+    end
+
+    def stp_port_priority
+      config_get('interface', 'stp_port_priority', name: @name)
+    end
+
+    def stp_port_priority=(val)
+      config_set('interface', 'stp_port_priority', name: @name, pp: val)
+    end
+
+    def default_stp_port_priority
+      config_get_default('interface', 'stp_port_priority')
+    end
+
+    # Getter: Builds an array of mst cost commands currently
+    # on the device.
+    #   cli: spanning-tree mst 0,2-4,6,8-12 cost 1000
+    #        spanning-tree mst 4000-4020 cost 2568
+    # array: [['0,2-4,6,8-12', '1000'], ['4000-4020', '2568']]
+    #
+    def stp_mst_cost
+      config_get('interface', 'stp_mst_cost', name: @name)
+    end
+
+    def stp_mst_cost=(list)
+      config_set('interface', 'stp_mst_cost',
+                 name: @name, state: 'no', range: @smr,
+                 val: '') if list.empty?
+      set_range_based_params(list, 'stp_mst_cost')
+    end
+
+    def default_stp_mst_cost
+      config_get_default('interface', 'stp_mst_cost')
+    end
+
+    # Getter: Builds an array of mst port-priority commands
+    # currently on the device.
+    #   cli: spanning-tree mst 0,2-4,6,8-12 port-priority 64
+    #        spanning-tree mst 4000-4020 port-priority 160
+    # array: [['0,2-4,6,8-12', '64'], ['4000-4020', '160']]
+    #
+    def stp_mst_port_priority
+      config_get('interface', 'stp_mst_port_priority', name: @name)
+    end
+
+    def stp_mst_port_priority=(list)
+      config_set('interface', 'stp_mst_port_priority',
+                 name: @name, state: 'no', range: @smr,
+                 val: '') if list.empty?
+      set_range_based_params(list, 'stp_mst_port_priority')
+    end
+
+    def default_stp_mst_port_priority
+      config_get_default('interface', 'stp_mst_port_priority')
+    end
+
+    def stp_port_type
+      config_get('interface', 'stp_port_type', name: @name)
+    end
+
+    def stp_port_type=(val)
+      if val
+        state = ''
+      else
+        state = 'no'
+        val = ''
+      end
+      config_set('interface', 'stp_port_type', name: @name,
+                 state: state, type: val)
+    end
+
+    def default_stp_port_type
+      config_get_default('interface', 'stp_port_type')
+    end
+
+    # Getter: Builds an array of vlan cost commands currently
+    # on the device.
+    #   cli: spanning-tree vlan 1-4,6,8-12 cost 1000
+    #        spanning-tree vlan 3000-3960 cost 2568
+    # array: [['1-4,6,8-12', '1000'], ['3000-3960', '2568']]
+    #
+    def stp_vlan_cost
+      config_get('interface', 'stp_vlan_cost', name: @name)
+    end
+
+    def stp_vlan_cost=(list)
+      config_set('interface', 'stp_vlan_cost',
+                 name: @name, state: 'no',
+                 range: @svr, val: '') if list.empty?
+      set_range_based_params(list, 'stp_vlan_cost')
+    end
+
+    def default_stp_vlan_cost
+      config_get_default('interface', 'stp_vlan_cost')
+    end
+
+    # Getter: Builds an array of vlan port-priority commands
+    # currently on the device.
+    #   cli: spanning-tree vlan 1-4,6,8-12 port-priority 64
+    #        spanning-tree vlan 3000-3960 port-priority 160
+    # array: [['1-4,6,8-12', '64'], ['3000-3960', '160']]
+    #
+    def stp_vlan_port_priority
+      config_get('interface', 'stp_vlan_port_priority', name: @name)
+    end
+
+    def stp_vlan_port_priority=(list)
+      config_set('interface', 'stp_vlan_port_priority',
+                 name: @name, state: 'no',
+                 range: @svr, val: '') if list.empty?
+      set_range_based_params(list, 'stp_vlan_port_priority')
+    end
+
+    def default_stp_vlan_port_priority
+      config_get_default('interface', 'stp_vlan_port_priority')
+    end
+
     def switchport
       # This is "switchport", not "switchport mode"
-      config_get('interface', 'switchport', @name)
+      config_get('interface', 'switchport', name: @name)
     end
 
     def switchport_enable(val=true)
-      config_set('interface', 'switchport', @name, val ? '' : 'no')
+      config_set('interface', 'switchport', name: @name, state: val ? '' : 'no')
     end
 
     # switchport_autostate_exclude is exclusive to switchport interfaces
     def switchport_autostate_exclude
       config_get('interface',
-                 'switchport_autostate_exclude', @name)
+                 'switchport_autostate_exclude', name: @name)
     end
 
     def switchport_autostate_exclude=(val)
-      # cannot configure autostate unless feature vlan is enabled
-      fail('switchport mode must be configured before ' \
-           'switchport autostate') unless switchport
-      feature_vlan_set(true)
+      if platform == :nexus
+        # cannot configure autostate unless feature vlan is enabled
+        fail('switchport mode must be configured before ' \
+             'switchport autostate') unless switchport
+        feature_vlan_set(true)
+      end
       config_set('interface', 'switchport_autostate_exclude',
-                 @name, val ? '' : 'no')
+                 name: @name, state: val ? '' : 'no')
     end
 
     def default_switchport_autostate_exclude
@@ -571,9 +809,9 @@ module Cisco
 
     def switchport_mode_lookup_string
       case @name
-      when /Ethernet/i
+      when ETHERNET
         return 'switchport_mode_ethernet'
-      when /port-channel/i
+      when PORTCHANNEL
         return 'switchport_mode_port_channel'
       else
         return 'switchport_mode_other_interfaces'
@@ -581,7 +819,8 @@ module Cisco
     end
 
     def switchport_mode
-      mode = config_get('interface', switchport_mode_lookup_string, @name)
+      return nil if platform == :ios_xr
+      mode = config_get('interface', switchport_mode_lookup_string, name: @name)
 
       return mode.nil? ? :disabled : IF_SWITCHPORT_MODE.key(mode)
 
@@ -589,6 +828,7 @@ module Cisco
       # Assume this is an interface that doesn't support switchport.
       # Do not raise exception since the providers will prefetch this property
       # regardless of interface type.
+      # TODO: this should probably be nil instead
       return :disabled
     end
 
@@ -600,11 +840,8 @@ module Cisco
       elsif :fex_fabric == mode_set
         fex_feature_set(:enabled) unless :enabled == fex_feature
       end
-      config_set('interface', switchport_mode_lookup_string, @name, '',
-                 IF_SWITCHPORT_MODE[mode_set])
-
-    rescue RuntimeError
-      raise "[#{@name}] switchport_mode is not supported on this interface"
+      config_set('interface', switchport_mode_lookup_string,
+                 name: @name, state: '', mode: IF_SWITCHPORT_MODE[mode_set])
     end
 
     def switchport_mode=(mode_set)
@@ -617,13 +854,13 @@ module Cisco
       when :disabled
         if switchport
           # Note: turn off switchport command, not switchport mode
-          config_set('interface', 'switchport', @name, 'no')
+          config_set('interface', 'switchport', name: @name, state: 'no')
         end
 
       when :default
         if :disabled == default_switchport_mode
           config_set('interface', switchport_mode_lookup_string,
-                     @name, 'no', '')
+                     name: @name, state: 'no', mode: '')
         else
           switchport_enable_and_mode(mode_set)
         end
@@ -631,31 +868,27 @@ module Cisco
       else
         switchport_enable_and_mode(mode_set)
       end # case
-
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
     end
 
     def default_switchport_mode
+      return nil if platform == :ios_xr
       return :disabled unless system_default_switchport
       IF_SWITCHPORT_MODE.key(
         config_get_default('interface', switchport_mode_lookup_string))
     end
 
     def switchport_trunk_allowed_vlan
-      config_get('interface', 'switchport_trunk_allowed_vlan', @name)
+      config_get('interface', 'switchport_trunk_allowed_vlan', name: @name)
     end
 
     def switchport_trunk_allowed_vlan=(val)
       if val.nil?
-        config_set(
-          'interface', 'switchport_trunk_allowed_vlan', @name, 'no', '')
+        config_set('interface', 'switchport_trunk_allowed_vlan',
+                   name: @name, state: 'no', vlan: '')
       else
-        config_set(
-          'interface', 'switchport_trunk_allowed_vlan', @name, '', val)
+        config_set('interface', 'switchport_trunk_allowed_vlan',
+                   name: @name, state: '', vlan: val)
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
     end
 
     def default_switchport_trunk_allowed_vlan
@@ -663,19 +896,17 @@ module Cisco
     end
 
     def switchport_trunk_native_vlan
-      config_get('interface', 'switchport_trunk_native_vlan', @name)
+      config_get('interface', 'switchport_trunk_native_vlan', name: @name)
     end
 
     def switchport_trunk_native_vlan=(val)
       if val.nil?
-        config_set(
-          'interface', 'switchport_trunk_native_vlan', @name, 'no', '')
+        config_set('interface', 'switchport_trunk_native_vlan',
+                   name: @name, state: 'no', vlan: '')
       else
-        config_set(
-          'interface', 'switchport_trunk_native_vlan', @name, '', val)
+        config_set('interface', 'switchport_trunk_native_vlan',
+                   name: @name, state: '', vlan: val)
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
     end
 
     # vlan_mapping & vlan_mapping_enable
@@ -698,13 +929,13 @@ module Cisco
     end
 
     def vlan_mapping
-      match = config_get('interface', 'vlan_mapping', @name)
+      match = config_get('interface', 'vlan_mapping', name: @name)
       match.each(&:compact!) unless match.nil?
       match
     end
 
     def vlan_mapping=(should_list)
-      Vni.feature_vni_enable unless Vni.feature_vni_enabled
+      Feature.vni_enable
 
       # Process a hash of vlan_mapping cmds from delta_add_remove().
       # The vlan_mapping cli does not allow commands to be updated, they must
@@ -714,12 +945,12 @@ module Cisco
       return if delta_hash.values.flatten.empty?
       # Process :remove first to ensure "update" commands will not fail.
       [:remove, :add].each do |action|
-        CiscoLogger.debug("vlan_mapping delta #{@get_args}\n"\
-                          "#{action}: #{delta_hash[action]}")
+        Cisco::Logger.debug("vlan_mapping delta #{@get_args}\n"\
+                            "#{action}: #{delta_hash[action]}")
         delta_hash[action].each do |original, translated|
           state = (action == :add) ? '' : 'no'
-          config_set('interface', 'vlan_mapping', @name,
-                     state, original, translated)
+          config_set('interface', 'vlan_mapping', name: @name,
+                     state: state, original: original, translated: translated)
         end
       end
     rescue Cisco::CliError => e
@@ -732,12 +963,12 @@ module Cisco
     end
 
     def vlan_mapping_enable
-      config_get('interface', 'vlan_mapping_enable', @name)
+      config_get('interface', 'vlan_mapping_enable', name: @name)
     end
 
     def vlan_mapping_enable=(state)
-      config_set('interface', 'vlan_mapping_enable', @name,
-                 state ? '' : 'no')
+      config_set('interface', 'vlan_mapping_enable',
+                 name: @name, state: state ? '' : 'no')
     end
 
     def default_switchport_trunk_native_vlan
@@ -782,16 +1013,14 @@ module Cisco
     end
 
     def switchport_vtp
-      return false unless switchport_vtp_mode_capable?
-      config_get('interface', 'vtp', @name)
+      return nil unless switchport_vtp_mode_capable?
+      config_get('interface', 'vtp', name: @name)
     end
 
     def switchport_vtp=(vtp_set)
       return false unless switchport_vtp_mode_capable?
       no_cmd = (vtp_set) ? '' : 'no'
-      config_set('interface', 'vtp', @name, no_cmd)
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      config_set('interface', 'vtp', name: @name, state: no_cmd)
     end
 
     def svi_cmd_allowed?(cmd)
@@ -802,13 +1031,14 @@ module Cisco
     # svi_autostate is exclusive to svi interfaces
     def svi_autostate
       return nil unless @name[/^vlan/i]
-      config_get('interface', 'svi_autostate', @name)
+      config_get('interface', 'svi_autostate', name: @name)
     end
 
     def svi_autostate=(val)
       check_switchport_disabled
       svi_cmd_allowed?('autostate')
-      config_set('interface', 'svi_autostate', @name, val ? '' : 'no')
+      config_set('interface', 'svi_autostate',
+                 name: @name, state: val ? '' : 'no')
     end
 
     def default_svi_autostate
@@ -821,19 +1051,20 @@ module Cisco
 
     def feature_vlan_set(val)
       return if feature_vlan? == val
-      config_set('interface', 'feature_vlan', val ? '' : 'no')
+      config_set('interface', 'feature_vlan', state: val ? '' : 'no')
     end
 
     # svi_management is exclusive to svi interfaces
     def svi_management
       return nil unless @name[/^vlan/i]
-      config_get('interface', 'svi_management', @name)
+      config_get('interface', 'svi_management', name: @name)
     end
 
     def svi_management=(val)
       check_switchport_disabled
       svi_cmd_allowed?('management')
-      config_set('interface', 'svi_management', @name, val ? '' : 'no')
+      config_set('interface', 'svi_management',
+                 name: @name, state: val ? '' : 'no')
     end
 
     def default_svi_management
@@ -849,21 +1080,22 @@ module Cisco
     end
 
     def check_switchport_disabled
-      fail "#{caller[0][/`.*'/][1..-2]} cannot be set unless switchport mode" \
-        ' is disabled' unless switchport_mode == :disabled
+      return if switchport_mode == :disabled || switchport_mode.nil?
+      fail("#{caller[0][/`.*'/][1..-2]} cannot be set unless " \
+           'switchport mode is disabled')
     end
 
     def vpc_id
-      config_get('interface', 'vpc_id', @name)
+      config_get('interface', 'vpc_id', name: @name)
     end
 
     def vpc_id=(num)
       if num
-        config_set('interface', 'vpc_id', @name, '', num)
+        config_set('interface', 'vpc_id', name: @name, state: '', id: num)
       else
         # 'no vpc' doesn't work for phy ports, so do a get
         num = vpc_id
-        config_set('interface', 'vpc_id', @name, 'no', num)
+        config_set('interface', 'vpc_id', name: @name, state: 'no', id: num)
       end
     end
 
@@ -872,12 +1104,12 @@ module Cisco
     end
 
     def vpc_peer_link
-      config_get('interface', 'vpc_peer_link', @name)
+      config_get('interface', 'vpc_peer_link', name: @name)
     end
 
     def vpc_peer_link=(state)
       no_cmd = (state ? '' : 'no')
-      config_set('interface', 'vpc_peer_link', @name, no_cmd)
+      config_set('interface', 'vpc_peer_link', name: @name, state: no_cmd)
     end
 
     def default_vpc_peer_link
@@ -885,22 +1117,48 @@ module Cisco
     end
 
     def vrf
-      config_get('interface', 'vrf', @name)
+      config_get('interface', 'vrf', name: @name)
     end
 
-    def vrf=(vrf)
-      fail TypeError unless vrf.is_a?(String)
-      if vrf.empty?
-        config_set('interface', 'vrf', @name, 'no', '')
-      else
-        config_set('interface', 'vrf', @name, '', vrf)
+    def vrf=(v)
+      fail TypeError unless v.is_a?(String)
+      return if v == vrf
+      # Changing the VRF can result in loss of IP address, so cache it
+      addr_1 = ipv4_address
+      mask_1 = ipv4_netmask_length
+      addr_2 = ipv4_address_secondary
+      mask_2 = ipv4_netmask_length_secondary
+      # XR actually blocks you from changing the VRF if IP addr is present
+      unless platform == :nexus
+        ipv4_addr_mask_set(nil, nil, false) unless addr_1.nil?
+        ipv4_addr_mask_set(nil, nil, true) unless addr_2.nil?
       end
-    rescue Cisco::CliError => e
-      raise "[#{@name}] '#{e.command}' : #{e.clierror}"
+      if v.empty?
+        config_set('interface', 'vrf', name: @name, state: 'no', vrf: '')
+      else
+        config_set('interface', 'vrf', name: @name, state: '', vrf: v)
+      end
+      ipv4_addr_mask_set(addr_1, mask_1, false) unless addr_1.nil?
+      ipv4_addr_mask_set(addr_2, mask_2, true) unless addr_2.nil?
     end
 
     def default_vrf
       config_get_default('interface', 'vrf')
+    end
+
+    def set_range_based_params(list, param_name)
+      list.each do |range, property_value|
+        # if a particular range is set to default, use 'no' cmd
+        if property_value == 'default'
+          config_set('interface', param_name,
+                     name: @name, state: 'no',
+                     range: range, val: '')
+        else
+          config_set('interface', param_name,
+                     name: @name, state: '',
+                     range: range, val: property_value)
+        end
+      end
     end
   end  # Class
 end    # Module