+++ lib/cisco_node_utils/vlan.rb
@@ -82,30 +82,27 @@ module Cisco
       return default_mode if result.nil?
       case result
       when /fabricpath/i
-        return 'fabricpath'
+        return :fabricpath
       when /ce/i
-        return 'ce'
+        return :CE
       end
     end
 
     def mode=(str)
       str = str.to_s
       if str.empty?
-        result = config_set('vlan', 'mode', @vlan_id, 'no', '')
+        config_set('vlan', 'mode', @vlan_id, 'no', '')
       else
         if 'fabricpath' == str
           fabricpath_feature_set(:enabled) unless
             :enabled == fabricpath_feature
         end
-        result = config_set('vlan', 'mode', @vlan_id, '', str)
+        config_set('vlan', 'mode', @vlan_id, '', str)
       end
-      cli_error_check(result)
-    rescue CliError => e
-      raise "[vlan #{@vlan_id}] '#{e.command}' : #{e.clierror}"
     end
 
     def default_mode
-      config_get_default('vlan', 'mode')
+      config_get_default('vlan', 'mode').to_sym
     end
 
     def vlan_name