+++ tests/ciscotest.rb
@@ -31,7 +31,33 @@ class CiscoTestCase < TestCase
   @@interfaces_id = nil
   # rubocop:enable Style/ClassVars
 
-  def node
+  # The feature (lib/cisco_node_utils/cmd_ref/<feature>.yaml) that this
+  # test case is associated with, if applicable.
+  # If the YAML file excludes this entire feature for this platform
+  # (top-level _exclude statement, not individual attributes), then
+  # all tests in this test case will be skipped.
+  @skip_unless_supported = nil
+
+  class << self
+    attr_reader :skip_unless_supported
+  end
+
+  def self.runnable_methods
+    return super if skip_unless_supported.nil?
+    return super if node.cmd_ref.supports?(skip_unless_supported)
+    # If the entire feature under test is unsupported,
+    # undefine the setup/teardown methods (if any) and skip the whole test case
+    remove_method :setup if instance_methods(false).include?(:setup)
+    remove_method :teardown if instance_methods(false).include?(:teardown)
+    [:all_skipped]
+  end
+
+  def all_skipped
+    skip("Skipping #{self.class}; feature " \
+         "'#{self.class.skip_unless_supported}' is unsupported on this node")
+  end
+
+  def self.node
     unless @@node
       @@node = Node.instance # rubocop:disable Style/ClassVars
       @@node.connect(address, username, password)
@@ -41,13 +67,17 @@ class CiscoTestCase < TestCase
       puts "\nNode under test:"
       puts "  - name  - #{@@node.host_name}"
       puts "  - type  - #{@@node.product_id}"
-      puts "  - image - #{@@node.system}\n\n"
+      # puts "  - image - #{@@node.system}\n\n"
     end
     @@node
-  rescue CiscoNxapi::HTTPUnauthorized
+  rescue Cisco::Client::AuthenticationFailed
     abort "Unauthorized to connect as #{username}:#{password}@#{address}"
-  rescue StandardError => e
+  rescue Cisco::Client::ClientError, TypeError, ArgumentError => e
     abort "Error in establishing connection: #{e}"
+  end
+
+  def node
+    self.class.node
   end
 
   def setup
@@ -59,8 +89,16 @@ class CiscoTestCase < TestCase
     node.cmd_ref
   end
 
+  def platform
+    node.client.platform
+  end
+
   def config(*args)
-    result = super
+    if node.client.platform == :ios_xr
+      result = super(*args, 'commit best-effort')
+    else
+      result = super
+    end
     node.cache_flush
     result
   end
@@ -81,7 +119,19 @@ class CiscoTestCase < TestCase
     # Compare the interface address with the current session address.
     # and return true if they match.
     return false if int_ip.nil?
-    int_ip == convert_dns_name(address)
+    int_ip == convert_dns_name(address.split(':')[0])
+  end
+
+  # Some NXOS hardware is not capable of supporting certain features even
+  # though the platform family in general includes support. In these cases
+  # the NU feature setter will raise a RuntimeError.
+  def hardware_supports_feature?(message)
+    patterns = ['Hardware is not capable of supporting',
+                'is unsupported on this node',
+               ]
+    skip('Skip test: Feature is unsupported on this device') if
+      message[Regexp.union(patterns)]
+    flunk(message)
   end
 
   def interfaces